--- SpeckleToolkit/Speckle_Adapter/CRUD/Read.cs
@@ -1,71 +1,0 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using BH.oM.Base;
-
-namespace BH.Adapter.Speckle
-{
-  public partial class SpeckleAdapter
-  {
-        /***************************************************/
-        /**** Adapter overload method                   ****/
-        /***************************************************/
-        protected override IEnumerable<IBHoMObject> Read(Type type, IList ids)
-        {
-            ////Choose what to pull out depending on the type. Also see example methods below for pulling out bars and dependencies
-            //if (type == typeof(Node))
-            //    return ReadNodes(ids as dynamic);
-            //else if (type == typeof(Bar))
-            //    return ReadBars(ids as dynamic);
-            //else if (type == typeof(ISectionProperty) || type.GetInterfaces().Contains(typeof(ISectionProperty)))
-            //    return ReadSectionProperties(ids as dynamic);
-            //else if (type == typeof(Material))
-            //    return ReadMaterials(ids as dynamic);
-
-            return new List<IBHoMObject>();
-        }
-
-        /***************************************************/
-        /**** Private specific read methods             ****/
-        ///***************************************************/
-
-        ////The List<string> in the methods below can be changed to a list of any type of identification more suitable for the toolkit
-
-        //private List<Bar> ReadBars( List<string> ids = null )
-        //{
-        //  //Implement code for reading bars
-        //  throw new NotImplementedException();
-        //}
-
-        ///***************************************/
-
-        //private List<Node> ReadNodes( List<string> ids = null )
-        //{
-        //  //Implement code for reading nodes
-        //  throw new NotImplementedException();
-        //}
-
-        ///***************************************/
-
-        //private List<ISectionProperty> ReadSectionProperties( List<string> ids = null )
-        //{
-        //  //Implement code for reading section properties
-        //  throw new NotImplementedException();
-        //}
-
-        ///***************************************/
-
-        //private List<Material> ReadMaterials( List<string> ids = null )
-        //{
-        //  //Implement code for reading materials
-        //  throw new NotImplementedException();
-        //}
-
-        /***************************************************/
-
-
-    }
-}
