--- SpeckleToolkit/Speckle_Engine/Convert/AdapterId.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace BH.Engine.Speckle
-{
-  public static partial class Convert
-  {
-    public const string AdapterId = "Speckle_id";
-  }
-}
