--- SpeckleToolkit/Speckle_Engine/Convert/ToSpeckle.cs
@@ -1,24 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace BH.Engine.Speckle
-{
-  public static partial class Convert
-  {
-    /***************************************************/
-    /**** Public Methods                            ****/
-    /***************************************************/
-
-    //Add methods for converting From BHoM to the specific software types, if possible to do without any BHoM calls
-    //Example:
-    //public static SpeckleNode ToSpeckle(this Node node)
-    //{
-    //    //Insert code for convertion
-    //}
-
-    /***************************************************/
-  }
-}
